libtrap: bugfix python sender won't exit without any client
When no receiver is connected, sender written in python (e.g.
debug_sender) won't exit even though SIGINT was already catched.
The reason is that output IFC waits at semphore for any client.

The solution is based on replacement of sem_wait() using
sem_timedwait().  When TRAP_WAIT is used, there is 1 second for waiting
and after that loop repeats.

In addition, to break blocking select() call, new pair of file
descriptors (pipe()) was added.  terminate() closes one end of pipe that
should be detected on the other file descriptor.

The problem should be fixed.