b_plus_tree: compare func may return any positive/negative number
Previously, compare func must return 0, 1 or -1.
This is problem since standard functions like memcmp or strcmp may return any negative or positive value. It is not guaranteed to be +1 or -1.

Changed, so compare func may now return any positive or negative value, so it is
comaptible with memcmp and similar.