libtrap: bugfix in TCPIP interface
After a socket was closed because of unsuccessful negotiation,
it was not set to -1. This caused other functions to treat the socket as valid.
Such bahevior resulted in situation where this socket was passed to select() function.
Select() function returns EBADF when a bad file descriptor is passed.
Further fucntions subsequently segfaulted.