report2idea: take `namespace` from config file
In production, we need to set a name of instance that is registered in
Warden server (in case we want to report alerts).  It was necessary to
set such name to each repoter as an argument.  This patch allows us to
set the common prefix of the name (such as cz.cesnet.nemea) in the
config file:

```yaml
namespace: cz.cesnet.nemea
```

and each reporter adds its module_name automatically, so we get
`cz.cesnet.nemea.blacklistfilter` for `blacklistfilter`.

According to feedback by @vaclavbartos, the `-n` argument was is still
supported.