treewide: prefer Python3 instead of Python
When I compiled it successfully on Ubuntu 20.04 LTS,
there was the following advise:

--
For installation of Python components run the following commands as root:
(cd pytrap; python setup.py install --record=installed-files.txt;)
(cd pycommon; python setup.py install --record=installed-files.txt;)
--

This ends up in my case that I don't have installed setuptools, but it
was trying to install for Python2. When I used python3
setup.py, it works. Let's force users to use python3.

Ubuntu provides package python-is-python2 by default.
User can choose if he wants to have this or python-is-python3.
If you don't have symlink for python, it is better to use python3.
On the other hand, Arch Linux symlinks python to python3. On Gentoo,
the choise is up user. Commonly python should be linked to python3 in
some way.

More details are in PEP 394. Since, there is removed python2 support for
pytrap and pycommon and in shebangs you are using Python3, there should be python3
and in spec file there is explicit dependency on Python3 packages.

python should be used if it is possible to use Python2 or Python3.