unirec: Fix build with GCC v10.x
According to https://gcc.gnu.org/gcc-10/changes.html, new GCC has changed
default behaviour so we get compile errors without -fcommon.

This is not required for the unirec library itself, which builds fine,
but for (some of) the users in the modules repository.

The unirec usage depends heavily on the -fcommon behaviour and fixing
that would require major redesign, so just explicitly add this option
to keep the old GCC behaviour.

This should help with: https://github.com/CESNET/Nemea/issues/20