Revert "build: gcc10+ compatibility: add -fcommon if gcc version is at least 10"
Since the fix in commit 498778e146ed871b6a4e0b6996751d69e93e8b94
does not work on platforms where GCC's -dumpversion does not report
only the major version, I missed that when I got hit by the bug.

As the fix for libtrap to allow -fno-common was not hard and is
implemented in commit 3f0a120eb196dff78275829a71d793508bdcba59,
this can be safely reverted as it is not needed.